+++ CHANGELOG.md
@@ -7,3 +7,10 @@ All notable changes to this project will be documented in this file. See [standa
 ### Features
 
 - Add Genesys PureCloud integration example ([1bb6d65](https://github.com/watson-developer-cloud/assistant-web-chat-service-desk-starter/commit/1bb6d6556df7429f2a8cd2203a5b641bb38b7751))
+
+## 2.0.0 (2020-10-22)
+
+### Features
+
+- Add Twilio Flex integration example
+- BREAKING CHANGE: Make buildEntry getInstance environment variable driven